abstract focusHandler into a separate hook
